+++ modules/Food_Service/Users/ServeMenus.php
@@ -44,6 +44,8 @@ if ( $_REQUEST['modfunc'] === 'submit' )
 			AND fsmi.ITEM_ID=fsi.ITEM_ID
 			AND fsmi.MENU_ID='" . (int) $_REQUEST['menu_id'] . "'", [], [ 'SHORT_NAME' ] );
 
+		$item_id = 0;
+
 		foreach ( (array) $_SESSION['FSA_sale'] as $item_sn )
 		{
 			$price = $items_RET[$item_sn][1]['PRICE_STAFF'];
@@ -51,8 +53,9 @@ if ( $_REQUEST['modfunc'] === 'submit' )
 			DBInsert(
 				'food_service_staff_transaction_items',
 				[
-					// @since 11.2 FS transaction item ID references food_service_menu_items(menu_item_id)
-					'ITEM_ID' => (int) $items_RET[$item_sn][1]['MENU_ITEM_ID'],
+					'ITEM_ID' => $item_id++,
+					// @since 11.2.1 FS transaction item ID references food_service_menu_items(menu_item_id)
+					'MENU_ITEM_ID' => (int) $items_RET[$item_sn][1]['MENU_ITEM_ID'],
 					'TRANSACTION_ID' => (int) $transaction_id,
 					'AMOUNT' => '-' . $price,
 					'SHORT_NAME' => DBEscapeString( $items_RET[$item_sn][1]['SHORT_NAME'] ),